+++ src/controller.ts
@@ -10,6 +10,7 @@ import * as utils from './utils';
 import { workspace } from "./workspace";
 import { Constants } from "./constants";
 import { bodyPaneContextMenuData, menuData, outlinePaneContextMenuData } from "./menu";
+import { toolbarButtons } from "./toolbar-buttons";
 import { keybindings } from "./keybindings";
 import { QuickSearchController } from "./core/quicksearch";
 import { nullButtonWidget } from "./core/leoFrame";
@@ -35,6 +36,7 @@ export class Controller {
 
     constructor() {
         workspace.menu.buildMenu(menuData);
+        workspace.menu.buildIconButtons(toolbarButtons);
         workspace.menu.buildBodyContextMenu(bodyPaneContextMenuData);
         workspace.menu.buildOutlineContextMenu(outlinePaneContextMenuData);
         workspace.layout.setWindowTitle(Constants.DEFAULT_WINDOW_TITLE);
@@ -78,6 +80,7 @@ export class Controller {
         this.setupButtonHandlers();
         this.setupConfigCheckboxes();
         this.setupTopMenuHandlers();
+        this.setupToolbarHandlers();
     }
     //@+node:felix.20260322222009.1: *4* setupOutlinePaneHandlers
     private setupOutlinePaneHandlers() {
@@ -446,6 +449,18 @@ export class Controller {
         });
 
     }
+    //@+node:felix.20260408001315.1: *4* setupToolbarHandlers
+    private setupToolbarHandlers() {
+        // setup horizontal mouse wheel scrolling for toolbar
+        const TOOLBAR = workspace.menu.TOOLBAR;
+        TOOLBAR.addEventListener('wheel', (e) => {
+            if (e.deltaY !== 0) {
+                e.preventDefault();
+            }
+            TOOLBAR.scrollLeft += e.deltaY * 1.5; // scroll horizontally, adjust multiplier as needed.
+        }, { passive: false });
+    }
+
     //@+node:felix.20260322221701.1: *4* setupDocumentTabsAndHandlers
     public setupDocumentTabsAndHandlers() {
         // The opened documents are in g.app.windowList.