+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.0.9
 
-- ...
+- Made button-toolbar mouse-scrollable horizontally.
+- Added icon buttons for most common commands in toolbar.
 
 ## 1.0.8